+++ src/gentables/fluid_interp_coeff_linear.c
@@ -10,7 +10,7 @@ const fluid_real_t interp_coeff_linear[] =
 //    interp_coeff_linear[i][0] = (1.0 - x);
 //    interp_coeff_linear[i][1] = x;
 //}
-#define x(i) ((double)(i) / (double)FLUID_INTERP_MAX)
+#define x(i) ((i) / (double)FLUID_INTERP_MAX)
 #define X(i) (fluid_real_t)(1.0 - x(i)), (fluid_real_t)x(i),
 #define AUTO_GEN_ARRAY_SIZE FLUID_INTERP_MAX
 #include "auto_gen_array.h"
